Lan tiep

+++ DoAn3/admin/modules/schools_index.php
@@ -1,6 +1,7 @@
 
 <?php include '../autoload/autoload.php' ;
 $school=$db->fetchAll("schools");
+$stt=$db->fetchAll("statuses");
 
 ?>
 
@@ -19,73 +20,7 @@ $school=$db->fetchAll("schools");
       <li class="breadcrumb-item active">Khoa viện</li>
     </ol>
 
-    <!-- Icon Cards-->
-    <div class="row">
-      <div class="col-xl-3 col-sm-6 mb-3">
-        <div class="card text-white bg-primary o-hidden h-100">
-          <div class="card-body">
-            <div class="card-body-icon">
-              <i class="fas fa-fw fa-comments"></i>
-            </div>
-            <div class="mr-5">26 New Messages!</div>
-          </div>
-          <a class="card-footer text-white clearfix small z-1" href="#">
-            <span class="float-left">View Details</span>
-            <span class="float-right">
-              <i class="fas fa-angle-right"></i>
-            </span>
-          </a>
-        </div>
-      </div>
-      <div class="col-xl-3 col-sm-6 mb-3">
-        <div class="card text-white bg-warning o-hidden h-100">
-          <div class="card-body">
-            <div class="card-body-icon">
-              <i class="fas fa-fw fa-list"></i>
-            </div>
-            <div class="mr-5">11 New Tasks!</div>
-          </div>
-          <a class="card-footer text-white clearfix small z-1" href="#">
-            <span class="float-left">View Details</span>
-            <span class="float-right">
-              <i class="fas fa-angle-right"></i>
-            </span>
-          </a>
-        </div>
-      </div>
-      <div class="col-xl-3 col-sm-6 mb-3">
-        <div class="card text-white bg-success o-hidden h-100">
-          <div class="card-body">
-            <div class="card-body-icon">
-              <i class="fas fa-fw fa-shopping-cart"></i>
-            </div>
-            <div class="mr-5">123 New Orders!</div>
-          </div>
-          <a class="card-footer text-white clearfix small z-1" href="#">
-            <span class="float-left">View Details</span>
-            <span class="float-right">
-              <i class="fas fa-angle-right"></i>
-            </span>
-          </a>
-        </div>
-      </div>
-      <div class="col-xl-3 col-sm-6 mb-3">
-        <div class="card text-white bg-danger o-hidden h-100">
-          <div class="card-body">
-            <div class="card-body-icon">
-              <i class="fas fa-fw fa-life-ring"></i>
-            </div>
-            <div class="mr-5">13 New Tickets!</div>
-          </div>
-          <a class="card-footer text-white clearfix small z-1" href="#">
-            <span class="float-left">View Details</span>
-            <span class="float-right">
-              <i class="fas fa-angle-right"></i>
-            </span>
-          </a>
-        </div>
-      </div>
-    </div>
+    <?php include '../layouts/info.php';?>
 
     <!-- Area Chart Example-->
     <div class="card mb-3">
@@ -106,41 +41,49 @@ $school=$db->fetchAll("schools");
            </div>
          <?php } ?>
        </div>
-         <div class="class row">
-          <div class="col-md-12">
-            <div class="table-responsive">
-              <table class="table table-bordered table-hover">
-                <thead>
-                  <tr>
-                    <th scope="col" style="text-align: center">ID</th>
-                    <th scope="col" style="text-align: center">Tên viện</th>
-                  </tr>
-                </thead>
-                <tbody>
-                  <?php foreach ($school as $item): ?>
-                    <tr>
-                      <td><?php echo $item['id'] ?></td>
-                      <td><?php echo $item['name'] ?></td>
-                       <td>
-                        <a class="btn btn-xs btn-info" href="schools_update.php?id= <?php echo $item['id']  ?>"><i class="fas fa-plus"></i>Sửa</a>
-                        <a class="btn btn-xs btn-danger" href="schools_delete.php?id= <?php echo $item['id']  ?>"><i class="fas fa-times"></i>Xóa</a>
+       <div class="class row">
+        <div class="col-md-12">
+          <div class="table-responsive">
+            <table class="table table-bordered table-hover">
+              <thead>
+                <tr>
+                  <th scope="col" style="text-align: center">ID</th>
+                  <th scope="col" style="text-align: center">Tên viện</th>
+                  <th scope="col" style="text-align: center">Trạng thái</th>
 
-                      </td>
-                    </tr>
+                </tr>
+              </thead>
+              <tbody>
+                <?php foreach ($school as $item): ?>
+                  <?php foreach ($stt as $item1): ?>
+                    <?php if ($item['state']==$item1['id']): ?>
+
+                      <tr>
+                        <td><?php echo $item['id'] ?></td>
+                        <td><?php echo $item['name'] ?></td>
+                        <td><?php echo $item1['status'] ?></td>
+                        <td>
+                          <a class="btn btn-xs btn-info" href="schools_update.php?id= <?php echo $item['id']  ?>"><i class="fas fa-plus"></i>Sửa</a>
+                          <a class="btn btn-xs btn-danger" href="schools_delete.php?id= <?php echo $item['id']  ?>"><i class="fas fa-times"></i>Xóa</a>
+
+                        </td>
+                      </tr>
+                    <?php endif ?>
                   <?php endforeach ?>
+                <?php endforeach ?>
 
-                </tbody>
-              </table>
-            </div>  
-          </div>
+              </tbody>
+            </table>
+          </div>  
         </div>
-
-        <div class="card-body">
-          <canvas id="myAreaChart" width="100%" height="30"></canvas>
-        </div>
-        <div class="card-footer small text-muted">Updated yesterday at 11:59 PM</div>
       </div>
 
+      <div class="card-body">
+        <canvas id="myAreaChart" width="100%" height="30"></canvas>
+      </div>
+      <div class="card-footer small text-muted">Updated yesterday at 11:59 PM</div>
+    </div>
 
-      <!-- /.container-fluid -->
-      <?php include '../layouts/footer.php' ?>
+
+    <!-- /.container-fluid -->
+    <?php include '../layouts/footer.php' ?>
